Push the last of the valid SPV logic into the SPV class
Minimal change to keep the diff small.

The service still calls SPV methods at innapropriate times however
this allows us to start simplifying from within the SPV class then
push its simplified state machine outwards.

Ideally external classes will merely notify the SPV manager that
things such as connectivity and login state have changed, and it
will choose how to respond internally.
